FullStack Application
First attempt at making a fullstack application to get some knowledge into how get/post works ,and how to use Spring boot, for java, and a bit of react for some frontend. I was following a guide, so it is not "selfmade" but i wrote it all myself, did some research, used copilot to explain stuff, and tried to understand it. This way of making applications is advanced for me, but omg is it a cool way to store data in a database, using spring boot, jpa and stuff like that